requirements.txt
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.0
googleapis-common-protos==1.56.4
httplib2==0.21.0
cachetools==5.2.0
protobuf===4.21.9
rsa==4.9
pynput===1.7.6